+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v3.1.19
+* replace stat2 by stat64
+* replace readdir2 by readdir64
+
 ## v3.1.18 (2022-06-23)
 * improve adbKit CLI